+++ coxeter/shapes/circle.py
@@ -47,7 +47,7 @@ class Circle(Shape2D):
 
     @property
     def gsd_shape_spec(self):
-        """dict: Get a :ref:`complete GSD specification <shapes>`."""  # noqa: D401
+        """dict: Get a :ref:`complete GSD specification <gsd:shapes>`."""  # noqa: D401
         return {"type": "Sphere", "diameter": 2 * self.radius}
 
     @property
@@ -220,4 +220,11 @@ class Circle(Shape2D):
         return (
             f"coxeter.shapes.Circle(radius={self.radius}, "
             f"center={self.centroid.tolist()})"
+        )
+
+    def _plato_primitive(self, backend):
+        return backend.Disks(
+            positions=np.array([self.center[:2]]),
+            colors=np.array([[0.5, 0.5, 0.5, 1]]),
+            radii=[self.radius],
         )