+++ coxeter/shapes/ellipsoid.py
@@ -61,7 +61,7 @@ class Ellipsoid(Shape3D):
 
     @property
     def gsd_shape_spec(self):
-        """dict: Get a :ref:`complete GSD specification <shapes>`."""  # noqa: D401
+        """dict: Get a :ref:`complete GSD specification <gsd:shapes>`."""  # noqa: D401
         return {"type": "Ellipsoid", "a": self.a, "b": self.b, "c": self.c}
 
     @property