+++ coxeter/shapes/polygon.py
@@ -183,7 +183,7 @@ class Polygon(Shape2D):
 
     @property
     def gsd_shape_spec(self):
-        """dict: Get a :ref:`complete GSD specification <shapes>`."""  # noqa: D401
+        """dict: Get a :ref:`complete GSD specification <gsd:shapes>`."""  # noqa: D401
         return {"type": "Polygon", "vertices": self.vertices.tolist()}
 
     @property
@@ -709,4 +709,13 @@ class Polygon(Shape2D):
         return (
             f"coxeter.shapes.Polygon(vertices={self.vertices.tolist()}, "
             f"normal={self.normal.tolist()})"
+        )
+
+    def _plato_primitive(self, backend):
+        verts = _align_points_by_normal(self.normal, self.vertices - self.center)[0]
+        return backend.Polygons(
+            positions=np.array([[0.0, 0.0]]),
+            orientations=np.array([[1.0, 0.0, 0.0, 0.0]]),
+            colors=np.array([[0.5, 0.5, 0.5, 1]]),
+            vertices=verts[:, :2],
         )