+++ coxeter/shapes/polyhedron.py
@@ -180,7 +180,7 @@ class Polyhedron(Shape3D):
 
     @property
     def gsd_shape_spec(self):
-        """dict: Get a :ref:`complete GSD specification <shapes>`."""  # noqa: D401
+        """dict: Get a :ref:`complete GSD specification <gsd:shapes>`."""  # noqa: D401
         return {
             "type": "Mesh",
             "vertices": self.vertices.tolist(),
@@ -857,4 +857,14 @@ class Polyhedron(Shape3D):
         return (
             f"coxeter.shapes.Polyhedron(vertices={self.vertices.tolist()}, "
             f"faces={np.asarray(self.faces).tolist()})"
+        )
+
+    def _plato_primitive(self, backend):
+        return backend.Mesh(
+            positions=np.array([self.center]),
+            orientations=np.array([[1.0, 0.0, 0.0, 0.0]]),
+            colors=np.array([[0.5, 0.5, 0.5, 1]] * len(self.vertices)),
+            vertices=self.vertices,
+            indices=self.faces,
+            shape_colors=np.array([[0.5, 0.5, 0.5, 1]]),
         )